+++ Source/Widgets/CesiumWidget/CesiumWidget.js
@@ -271,7 +271,6 @@ function CesiumWidget(container, options) {
       mapProjection: options.mapProjection,
       orderIndependentTranslucency: options.orderIndependentTranslucency,
       scene3DOnly: defaultValue(options.scene3DOnly, false),
-      terrainExaggeration: options.terrainExaggeration,
       shadows: options.shadows,
       mapMode2D: options.mapMode2D,
       requestRenderMode: options.requestRenderMode,
@@ -292,13 +291,6 @@ function CesiumWidget(container, options) {
     var globe = options.globe;
     if (!defined(globe)) {
       globe = new Globe(ellipsoid);
-    }
-    if (defined(options.terrainExaggeration)) {
-      deprecationWarning(
-        "terrainExaggeration-removed",
-        "terrainExaggeration is now a property of Globe"
-      );
-      globe.terrainExaggeration = options.terrainExaggeration;
     }
     if (globe !== false) {
       scene.globe = globe;